+++ docs/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.3.0] Unreleased
+
+### Added
+- Added PsqlSnapshot class to create and manage snapshots of PostgreSQL tables.
+
+### Changed
+- SchemaVersioning has been updated and separated to dataframe and PostgreSql versioning classes to correctly detect added and removed columns for both.
+- compute_snapshot() method added. 
+- Added type hints.
+
 ## [v0.2.0] Unreleased
 
 ### Added